+++ src/concrete/ml/common/utils.py
@@ -9,6 +9,7 @@ from typing import Any, Callable, Dict, Iterable, List, Optional, Tuple, Union
 import numpy
 import onnx
 import torch
+from concrete.fhe import Exactness
 from concrete.fhe.dtypes import Integer
 from sklearn.base import is_classifier, is_regressor
 
@@ -605,3 +606,48 @@ def array_allclose_and_same_shape(
     assert isinstance(b, numpy.ndarray)
 
     return a.shape == b.shape and numpy.allclose(a, b, rtol, atol, equal_nan)
+
+
+def check_rounding_threshold(rounding_threshold_bits):
+    """Check and process the rounding_threshold_bits parameter.
+
+    Args:
+        rounding_threshold_bits (Union[None, int, Dict[str, Union[str, int]]]): Defines precision
+            rounding for model accumulators. Accepts None, an int, or a dict.
+            The dict can specify 'method' (fhe.Exactness.EXACT or fhe.Exactness.APPROXIMATE)
+            and 'n_bits' ('auto' or int)
+
+    Returns:
+        Dict[str, Union[str, int]]: Processed rounding_threshold_bits dictionary.
+
+    Raises:
+        NotImplementedError: If 'auto' rounding is specified but not implemented.
+        ValueError: If an invalid type or value is provided for rounding_threshold_bits.
+    """
+    n_bits_rounding: Union[None, str, int] = None
+    method: Exactness = Exactness.EXACT
+
+    # Only process if rounding_threshold_bits is not None
+    if rounding_threshold_bits is not None:
+        if isinstance(rounding_threshold_bits, int):
+            n_bits_rounding = rounding_threshold_bits
+        elif isinstance(rounding_threshold_bits, dict):
+            n_bits_rounding = rounding_threshold_bits.get("n_bits")
+            if n_bits_rounding == "auto":
+                raise NotImplementedError("Automatic rounding is not implemented yet.")
+            method = rounding_threshold_bits.get("method", method)
+            if not isinstance(method, Exactness):
+                method_str = method.upper()
+                if method_str in ["EXACT", "APPROXIMATE"]:
+                    method = Exactness[method_str]
+                else:
+                    raise ValueError(
+                        f"{method_str} is not a valid method. Must be one of EXACT, APPROXIMATE."
+                    )
+        else:
+            raise ValueError("Invalid type for rounding_threshold_bits. Must be int or dict.")
+
+        if n_bits_rounding is not None and not 2 <= n_bits_rounding <= 8:
+            raise ValueError("n_bits_rounding must be between 2 and 8 inclusive.")
+
+    return {"n_bits": n_bits_rounding, "method": method}
