+++ src/concrete/ml/torch/compile.py
@@ -11,13 +11,14 @@ import torch
 from brevitas.export.onnx.qonnx.manager import QONNXManager as BrevitasONNXManager
 from brevitas.nn.quant_layer import QuantInputOutputLayer as QNNMixingLayer
 from brevitas.nn.quant_layer import QuantNonLinearActLayer as QNNUnivariateLayer
-from concrete.fhe import Exactness, ParameterSelectionStrategy
+from concrete.fhe import ParameterSelectionStrategy
 from concrete.fhe.compilation.artifacts import DebugArtifacts
 from concrete.fhe.compilation.configuration import Configuration
 
 from ..common.debugging import assert_false, assert_true
 from ..common.utils import (
     MAX_BITWIDTH_BACKWARD_COMPATIBLE,
+    check_rounding_threshold,
     check_there_is_no_p_error_options_in_configuration,
     get_onnx_opset_version,
     manage_parameters_for_pbs_errors,
@@ -30,52 +31,6 @@ from . import NumpyModule
 
 Tensor = Union[torch.Tensor, numpy.ndarray]
 Dataset = Union[Tensor, Tuple[Tensor, ...]]
-
-
-def check_rounding_threshold(rounding_threshold_bits):
-    """Check and process the rounding_threshold_bits parameter.
-
-    Args:
-        rounding_threshold_bits (Union[None, int, Dict[str, Union[str, int]]]): Defines precision
-            rounding for model accumulators. Accepts None, an int, or a dict.
-            The dict can specify 'method' (fhe.Exactness.EXACT or fhe.Exactness.APPROXIMATE)
-            and 'n_bits' ('auto' or int)
-
-    Returns:
-        Dict[str, Union[str, int]]: Processed rounding_threshold_bits dictionary.
-
-    Raises:
-        NotImplementedError: If 'auto' rounding is specified but not implemented.
-        ValueError: If an invalid type or value is provided for rounding_threshold_bits.
-    """
-    n_bits_rounding: Union[None, str, int] = None
-    method: Exactness = Exactness.EXACT
-
-    # Only process if rounding_threshold_bits is not None
-    if rounding_threshold_bits is not None:
-        if isinstance(rounding_threshold_bits, int):
-            n_bits_rounding = rounding_threshold_bits
-        elif isinstance(rounding_threshold_bits, dict):
-            n_bits_rounding = rounding_threshold_bits.get("n_bits")
-            if n_bits_rounding == "auto":
-                raise NotImplementedError("Automatic rounding is not implemented yet.")
-            method = rounding_threshold_bits.get("method", method)
-            if not isinstance(method, Exactness):
-                method_str = method.upper()
-                if method_str in ["EXACT", "APPROXIMATE"]:
-                    method = Exactness[method_str]
-                else:
-                    raise ValueError(
-                        f"{method_str} is not a valid method. Must be one of EXACT, APPROXIMATE."
-                    )
-        else:
-            raise ValueError("Invalid type for rounding_threshold_bits. Must be int or dict.")
-
-        assert n_bits_rounding is not None, "n_bits_rounding cannot be None"
-        if not 2 <= n_bits_rounding <= 8:
-            raise ValueError("n_bits_rounding must be between 2 and 8 inclusive.")
-
-    return {"n_bits": n_bits_rounding, "method": method}
 
 
 def has_any_qnn_layers(torch_model: torch.nn.Module) -> bool:
@@ -144,6 +99,8 @@ def build_quantized_module(
     Returns:
         QuantizedModule: The resulting QuantizedModule.
     """
+    rounding_threshold_bits = check_rounding_threshold(rounding_threshold_bits)
+
     inputset_as_numpy_tuple = tuple(
         convert_torch_tensor_or_numpy_array_to_numpy_array(val) for val in to_tuple(torch_inputset)
     )
